feat: componentização da tela de home
